[game/server] UTIL_ScreenFade* APIs should use short for flags to prevent warnings

+++ game/server/EnvFade.cpp
@@ -72,7 +72,8 @@ void CEnvFade::Spawn( void )
 //-----------------------------------------------------------------------------
 void CEnvFade::InputFade( inputdata_t &inputdata )
 {
-	int fadeFlags = 0;
+	// dimhotepus: int -> short
+	short fadeFlags = 0;
 
 	if ( m_spawnflags & SF_FADE_IN )
 	{
